move placeholder to VisualizerModal
